repos/keepassxc/requirements.txt
python-dotenv~=1.0.1
pykeepass~=4.1.0.post1
requests~=2.32.3